refactor: 공통으로 사용되는 클래스 common 패키지로 이동
